Move clean_xml to abstract.
